+++ sensitivity_games/controller.py
@@ -11,28 +11,3 @@ class Controller(object):
         Virtual function in abstract Controller class
         '''
         raise NotImplementedError('Not Implemented!')
-
-
-class RegulatorController(Controller):
-    '''
-    '''
-    def __init__(self,
-                 dynamics,
-                 xf):
-        super().__init__(dynamics)
-        self.xf = xf
-
-
-class NeuralNetworkController(Controller):
-    '''
-    '''
-    def __init__(self,
-                 dynamics,
-                 model):
-        super().__init__(dynamics)
-        self.model = model
-
-    def get_control(self, x):
-        '''
-        '''
-        return self.model.forward(x)
