+++ sensitivity_games/regulator_controller.py
@@ -1,0 +1,11 @@
+from sensitivity_games.controller import Controller
+
+
+class RegulatorController(Controller):
+    '''
+    '''
+    def __init__(self,
+                 dynamics,
+                 xf):
+        super().__init__(dynamics)
+        self.xf = xf
